langchain-s/langchain-sse/requirements.txt
fastapi==0.116.2
uvicorn[standard]==0.32.0
langchain==1.2.14
langchain-core==1.2.26
langchain-community==0.4.1
langchain-openai==1.1.2
# langchain-openai==0.0.2
openai==2.31.0
python-dotenv==1.0.0
dashscope==1.23.7
requests==2.32.3
